requirements.txt
asgiref==3.3.1
astroid==2.5.3
colorama==0.4.4
Django==3.1.7
django-crispy-forms==1.11.2
django-etc==1.3.0
django-hitcount==1.3.3
django-resized==0.3.11
django-taggit==1.3.0
django-tinymce==3.3.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.2.0
pylint==2.7.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
